refactor: 기존 클래스의 패키지 명 변경으로 인한 테스트 코드 패키지명 변경 (`activitylog` -> `event`)
Moved test classes from activitylog to event package and updated references from ActivityCodes (String) to ActivityType (enum) for activity publishing and verification. Removed UserActivityEventListenerErrorTest as part of test suite cleanup.
