+++ src/test/java/com/eventitta/gamification/event/ActivityEventPublisherTest.java
@@ -1,12 +1,15 @@
-package com.eventitta.gamification.activitylog;
+package com.eventitta.gamification.event;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.springframework.context.ApplicationEventPublisher;
 
+import static com.eventitta.gamification.domain.ActivityType.CREATE_POST;
+import static com.eventitta.gamification.domain.ActivityType.DELETE_POST;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 
 @DisplayName("사용자 활동 이벤트 퍼블리셔의 동작을 검증하는 테스트")
 class ActivityEventPublisherTest {
@@ -17,7 +20,7 @@ class ActivityEventPublisherTest {
         ApplicationEventPublisher mockPublisher = mock(ApplicationEventPublisher.class);
         ActivityEventPublisher eventPublisher = new ActivityEventPublisher(mockPublisher);
 
-        eventPublisher.publish("CODE", 1L, 2L);
+        eventPublisher.publish(CREATE_POST, 1L, 2L);
 
         ArgumentCaptor<Object> captor = ArgumentCaptor.forClass(Object.class);
         verify(mockPublisher).publishEvent(captor.capture());
@@ -31,7 +34,7 @@ class ActivityEventPublisherTest {
         ApplicationEventPublisher mockPublisher = mock(ApplicationEventPublisher.class);
         ActivityEventPublisher eventPublisher = new ActivityEventPublisher(mockPublisher);
 
-        eventPublisher.publishRevoke("CODE", 1L, 2L);
+        eventPublisher.publishRevoke(DELETE_POST, 1L, 2L);
 
         ArgumentCaptor<Object> captor = ArgumentCaptor.forClass(Object.class);
         verify(mockPublisher).publishEvent(captor.capture());
